apps/workers/python/asr_worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
whisperx==3.1.1
torch==2.1.1
torchaudio==2.1.1
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
redis==5.0.1
nats-py==2.6.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
